requirements.txt
Flask==3.0.2
PyMuPDF==1.23.26
Pillow==10.2.0
flask-cors==4.0.0
Werkzeug==3.0.1
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5